implemented user specific crud operations

+++ controllers/contactController.js
@@ -50,12 +50,17 @@ const updateContact = asyncHandler(async (req, res) => {
         throw new Error("Contact not found");
     }
 
+    if (contact.user_id.toString() !== req.user_id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
         {new: true}
     );
-    res.status(200).json(updatedContact)
+    res.status(200).json(updatedContact);
 });
 
 //@desc Delete contact
@@ -67,6 +72,11 @@ const deleteContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
     };
+
+    if (contact.user_id.toString() !== req.user_id) {
+        res.status(403);
+        throw new Error("User don't have permission to delete other user contacts");
+    }
 
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(contact)
